Adjust RPI settings for cleaner and more supported boot
One of the goals of CirnOS is to be entirely hot swappable -- that
is, the system can be updated without rebooting. It made sense to
enable hot swapping for the HDMI display in the config.txt file so
a user can only connect a monitor when they need to see what their
rPi is doing.

The rainbow splash screen seen on boothas been removed. This was
bothering me for a long time, as it seemed out of place in this
operating system's boot process.

kernel.img has been renamed to just 'cirnos'. A common complaint
was that the name kernel.img didn't make sense since the README
file says that CirnOS does not have a kernel. A plain cirnos file
makes much more sense.

The default resolution in config.txt has been changed to 720p.
This should allow users to plug the rPi into any HDMI monitor
without config. It used to be 1600x900, which is not guaranteed on
a lot of screens.

Different interfaces like i2c have been enabled in config.txt to
pave the way for new drivers.